% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llamada_api_barakaldo.R
\name{llamada_api_url}
\alias{llamada_api_url}
\title{llamada api}
\usage{
llamada_api_url(url)
}
\arguments{
\item{url}{}
}
\value{
json
}
\description{
llamada api
}
\examples{
 llamada_api_url("https://www.barakaldo.eus/axis2/services/BarakaldoServices/actosFecha?response=application/json&fecha=2018-06-15&X-JWT-Assertion=")

}
